requirements.txt
alembic==1.5.5
Babel==2.9.0
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
postgres==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
python-dateutil==2.6.0
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
WTForms==2.3.3
